requirements.txt
flask==2.3.3
langchain==0.3.26
langchain-google-genai
google-generativeai
python-dotenv==1.0.0
requests==2.32.4
pydantic==2.11.7
typing-extensions==4.14.1 